Fixed crashes in SkeletonObjectEditor
-Removed or changed each _() to wxT() for properties names: Properties
names should not be translated, as a bad translation or a missing
translation could prevent wxWidgets to find a property ( leading to a
crash on my computer ).
-Put the property grid into a sizer. Removed the manual resize of the
grid ( was crashing the editor on my computer ).
-Removed default size for panels ( may be also a potential source of
crash as it can mess up wxWidgets layout system ). Added instead a new
default size for the editor.
-Corrected crash when clicking on "Add a bone" button when no bone is
selected.